@@ -4,29 +4,29 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.2+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   # Spinner
-  flutter_spinkit: ^4.1.2
+  flutter_spinkit: ^5.0.0
   # Fonts
-  google_fonts: ^1.1.1
+  google_fonts: ^2.1.0
   # Imaging
-  cached_network_image: ^2.2.0+1
+  cached_network_image: ^3.0.0
   # Firebase
   # ***---REQUIREMENT---***
-  firebase_core: ^0.7.0
-  firebase_analytics: ^7.0.1
-  firebase_crashlytics: ^0.4.0+1
-  firebase_auth: ^0.20.1
-  cloud_firestore: ^0.16.0+1
-  firebase_storage: ^7.0.0
-  firebase_performance: ^0.5.0+1
+  firebase_core: ^1.2.0
+  firebase_analytics: ^8.1.0
+  firebase_crashlytics: ^2.0.4
+  firebase_auth: ^1.3.0
+  cloud_firestore: ^2.2.1
+  firebase_storage: ^8.1.1
+  firebase_performance: ^0.7.0+3
   # Provider
-  provider: ^4.3.3
+  provider: ^5.0.0
 
   cupertino_icons: ^1.0.0
 
