@@ -1,10 +1,10 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 
 class FacilityServiceModel {
-  String facilityName;
-  String facilityReference;
-  List<dynamic> services;
-  List<dynamic> facilityPaymentModalities;
+  String? facilityName;
+  String? facilityReference;
+  List<dynamic>? services;
+  List<dynamic>? facilityPaymentModalities;
 
   FacilityServiceModel({
     this.facilityName,
@@ -14,7 +14,7 @@ class FacilityServiceModel {
   });
 
   factory FacilityServiceModel.fromFirestore(DocumentSnapshot doc) {
-    Map data = doc.data();
+    dynamic data = doc.data();
     return FacilityServiceModel(
       facilityName: data['facilityName'] ?? '',
       facilityReference: data['facilityReference'],
