@@ -1,7 +1,7 @@
 class SocialMediaModel {
-  String twitterHandle;
-  String facebookHandle;
-  String youtubeHandle;
+  String? twitterHandle;
+  String? facebookHandle;
+  String? youtubeHandle;
 
   SocialMediaModel({
     this.twitterHandle,
