@@ -2,15 +2,15 @@ import 'package:flutter/material.dart';
 
 class CircleButton extends StatelessWidget {
   final IconData icon;
-  final Color color;
-  final double iconSize;
-  final Function onPressed;
-  final String tooltip;
+  final Color? color;
+  final double? iconSize;
+  final VoidCallback onPressed;
+  final String? tooltip;
 
   const CircleButton({
-    Key key,
-    @required this.icon,
-    @required this.onPressed,
+    Key? key,
+    required this.icon,
+    required this.onPressed,
     this.color,
     this.iconSize,
     this.tooltip,
