@@ -2,14 +2,14 @@ import 'package:flutter/material.dart';
 
 class Responsive extends StatelessWidget {
   final Widget mobile;
-  final Widget tablet;
+  final Widget? tablet;
   final Widget desktop;
 
   Responsive({
-    Key key,
-    @required this.mobile,
+    Key? key,
+    required this.mobile,
     this.tablet,
-    @required this.desktop,
+    required this.desktop,
   }) : super(key: key);
 
   static bool isMobile(BuildContext context) =>
