@@ -7,13 +7,13 @@ export 'app_drawer.dart';
 export 'tab_widget.dart';
 export 'job_post_card.dart';
 export 'practitioner_card.dart';
-export 'store_item.dart';
 export 'responsive.dart';
 
 // Future Helpers
 export 'data_retrieval/practitioner_data_retrieval.dart';
 export 'data_retrieval/service_data_retrieval.dart';
 export 'data_retrieval/facility_data_retrieval.dart';
+export 'data_retrieval/store_data_retrieval.dart';
 
 // Global
 export 'global/global_loader.dart';
